requirements.txt
openai==0.27.0
questionary==1.10.0
python-dotenv==1.0.0

